Add generic module
