+++ frontend/src/modules/generic/store/constants.js
@@ -1,0 +1,1 @@
+export const INITIAL_PAGE_SIZE = 20
