+++ frontend/src/modules/generic/store/index.js
@@ -1,0 +1,13 @@
+import state from './state.js'
+import actions from './actions.js'
+import getters from './getters.js'
+import mutations from './mutations.js'
+
+export default {
+  namespaced: true,
+
+  state,
+  actions,
+  getters,
+  mutations
+}
