+++ frontend/src/shared/filter/components/filter-list-operator.vue
