+++ frontend/src/shared/filter/components/filter-list.vue
