+++ frontend/src/shared/filter/components/type/filter-type-boolean.vue
