+++ frontend/src/shared/filter/components/type/filter-type-search.vue
