+++ frontend/src/shared/filter/components/type/filter-type-select-group.vue
