+++ frontend/src/shared/filter/components/type/filter-type-select-multi.vue
