+++ frontend/src/shared/filter/helpers/operators.js
