--- frontend/src/shared/filters/format-date-filter.js
@@ -1,12 +1,0 @@
-import moment from 'moment'
-
-export default {
-  name: 'formatDate',
-  implementation(value) {
-    if (value) {
-      return moment(value).format('YYYY-MM-DD')
-    }
-
-    return null
-  }
-}
