--- frontend/src/shared/filters/format-datetime-filter.js
@@ -1,12 +1,0 @@
-import moment from 'moment'
-
-export default {
-  name: 'formatDatetime',
-  implementation(value) {
-    if (value) {
-      return moment(value).format('YYYY-MM-DD HH:mm')
-    }
-
-    return null
-  }
-}
